Function wavefront_obj::obj::parse

pub fn parse<S: AsRef<str>>(input: S) -> Result<ObjSet, ParseError>

Parses a wavefront .obj file, returning either the successfully parsed file, or an error. Support in this parser for the full file format is best-effort and realistically I will only end up supporting the subset of the file format which falls under the "things I see exported from blender" category.