Type Definition wavefront_obj::obj::VTNIndex

type VTNIndex = (VertexIndex, Option<TextureIndex>, Option<NormalIndex>);

An index into the vertex array, with an optional index into the texture array. This is used to define the corners of shapes which may or may not be textured.